pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[project]
name = "simulation_server"
version = "0.0.1"
requires-python = ">=3.9"
description = "ExaDigiT Simulation Server"
readme = "README.md"
# license = {file = "LICENSE.txt"}

dependencies = [
    "pydantic@git+https://github.com/pydantic/pydantic@e99bf8a#egg=pydantic",
    "pydantic-settings==2.1.0 ",
    "loguru==0.7.2",
    "SQLAlchemy==2.0.27",
    "pydruid==0.6.6",
]

[project.optional-dependencies]
server = [
    "fastapi==0.108.0",
    "gunicorn==21.2.0 ",
    "uvicorn==0.27.1",
]
simulation = [
    "matplotlib==3.7.2",
    "numpy==1.23.5",
    "rich==13.6.0",
    "fmpy==0.3.19",
    "pandas==2.0.3",
    "scipy==1.10.1",
    "pyarrow==15.0.0",
    "PyYAML==6.0.1",
    "ClusterShell==1.9.2", 
]
dev = [
    "pytest",
]

# [project.scripts]
# spam-cli = "spam:main_cli"

# [project.entry-points."spam.magical"]
# tomatoes = "spam:main_tomatoes"

 